Generalize solver-internal reset to all Newton solvers
Move the SolverBase.reset(flags=0) call into the base
_reset_solver_internals hook so any solver that implements the upstream
reset API (currently MuJoCo; VBD or others if they add one) is covered
automatically. SolverBase.reset is a documented no-op, so solvers
without an override are unaffected. MJWarp's override shrinks to the
CPU-MuJoCo gating and delegates to the base; Kamino opts out since its
FK delegate already performs the masked SolverKamino.reset with an
explicit ResetConfig.

Also move the regression test from test_articulation.py to
test/physics/test_newton_solver_reset.py: the articulation suite is
solver-agnostic and the test is MJWarp-specific.
