+++ source/isaaclab_newton/isaaclab_newton/physics/kamino_manager.py
@@ -173,6 +173,20 @@ class NewtonKaminoManager(NewtonManager):
             )
 
     @classmethod
+    def _reset_solver_internals(cls, world_mask: wp.array) -> None:
+        """Skip the base hook's generic solver reset.
+
+        The FK delegate (:meth:`_eval_fk_impl`) already performs the masked
+        :meth:`SolverKamino.reset` with an explicit
+        :class:`SolverKamino.ResetConfig` on both branches, so the base hook's
+        generic ``solver.reset(flags=0)`` would run a redundant second reset
+        for the same worlds at every boundary.
+
+        Args:
+            world_mask: Unused; accepted to match the base hook signature.
+        """
+
+    @classmethod
     def _build_solver(cls, model: Model, solver_cfg: KaminoSolverCfg) -> None:
         """Construct :class:`SolverKamino` and populate the base-class slots.
