+++ source/isaaclab_newton/test/assets/test_articulation.py
@@ -26,7 +26,7 @@ import warp as wp
 from isaaclab_newton.assets import Articulation
 from isaaclab_newton.physics import MJWarpSolverCfg, NewtonCfg
 from isaaclab_newton.physics import NewtonManager as SimulationManager
-from newton.solvers import SolverMuJoCo, SolverNotifyFlags
+from newton.solvers import SolverNotifyFlags
 
 import isaaclab.sim as sim_utils
 import isaaclab.utils.math as math_utils
@@ -2568,92 +2568,6 @@ def test_body_q_consistent_after_root_write(num_articulations, device, articulat
         assert diff < 0.01, (
             f"body_q was stale when collide() ran: diff={diff:.4f}m, jq={jq_root.tolist()}, bq={bq_root.tolist()}"
         )
-
-
-@pytest.mark.parametrize("device", ["cuda:0"])
-def test_env_reset_clears_selected_mjwarp_solver_internals(device):
-    """An env reset clears the flagged world's MuJoCo warm-start history and keeps the others.
-
-    Regression test for NaN values persisting across env reset: MJWarp warm-starts each solve
-    from ``qacc_warmstart``, so solver-internal buffers must be cleared for reset worlds at both
-    boundaries that consume the accumulated reset masks (``forward()`` and ``step()``), while the
-    authored joint state and the untouched worlds' history are preserved.
-    """
-    from unittest.mock import patch
-
-    sim_cfg = SimulationCfg(
-        dt=1 / 120,
-        physics=NewtonCfg(
-            solver_cfg=MJWarpSolverCfg(
-                njmax=20,
-                nconmax=20,
-                integrator="implicitfast",
-            ),
-            num_substeps=1,
-            use_cuda_graph=False,
-        ),
-    )
-    with build_simulation_context(sim_cfg=sim_cfg, device=device) as sim:
-        sim._app_control_on_stop_handle = None
-        articulation, _ = generate_articulation(
-            generate_articulation_cfg(articulation_type="single_joint_explicit"),
-            num_articulations=2,
-            device=device,
-        )
-        sim.reset()
-
-        solver = SimulationManager._solver
-        assert isinstance(solver, SolverMuJoCo)
-        warm_start = wp.to_torch(solver.mjw_data.qacc_warmstart)
-        assert warm_start.shape[0] == 2
-        assert warm_start.shape[1] > 0
-
-        env_ids = torch.tensor([0], dtype=torch.int32, device=device)
-        joint_pos = articulation.data.default_joint_pos.torch[:1].clone() + 0.25
-        joint_vel = torch.full_like(articulation.data.default_joint_vel.torch[:1], 0.5)
-        articulation.write_joint_state_to_sim_index(
-            position=joint_pos,
-            velocity=joint_vel,
-            env_ids=env_ids,
-        )
-
-        state = SimulationManager._state_0
-        joint_q_before = wp.to_torch(state.joint_q).clone()
-        joint_qd_before = wp.to_torch(state.joint_qd).clone()
-        warm_start[0].fill_(13.0)
-        warm_start[1].fill_(17.0)
-        wp.synchronize_device(device)
-
-        # The public, non-integrating forward boundary consumes the authored-write mask.
-        sim.forward()
-        wp.synchronize_device(device)
-
-        torch.testing.assert_close(wp.to_torch(state.joint_q), joint_q_before)
-        torch.testing.assert_close(wp.to_torch(state.joint_qd), joint_qd_before)
-        assert torch.count_nonzero(warm_start[0]).item() == 0
-        torch.testing.assert_close(warm_start[1], torch.full_like(warm_start[1], 17.0))
-
-        # The physics-step boundary also resets when no forward boundary consumed the mask first.
-        articulation.write_joint_state_to_sim_index(
-            position=joint_pos,
-            velocity=joint_vel,
-            env_ids=env_ids,
-        )
-        warm_start[0].fill_(23.0)
-        warm_start[1].fill_(29.0)
-        wp.synchronize_device(device)
-
-        with (
-            patch.object(SimulationManager, "_simulate_full", classmethod(lambda cls: None)),
-            patch.object(SimulationManager, "_simulate_physics_only", classmethod(lambda cls: None)),
-        ):
-            sim.step(render=False)
-        wp.synchronize_device(device)
-
-        torch.testing.assert_close(wp.to_torch(state.joint_q), joint_q_before)
-        torch.testing.assert_close(wp.to_torch(state.joint_qd), joint_qd_before)
-        assert torch.count_nonzero(warm_start[0]).item() == 0
-        torch.testing.assert_close(warm_start[1], torch.full_like(warm_start[1], 29.0))
 
 
 @pytest.mark.parametrize("add_ground_plane", [True])
